Add F-Droid build flavor (Closes #128) (#162)
* Move getSafeNameForSensor to StardroidApplication

It's only used there and allows use to make Analytics an interface
later.

* Cleanup build script

- Add -e to error on build failures.
- Use subshell instead of cd ..
- Manually fix classpath because gradle has problems with flavors

* Extract Analytics.java content into an interface

* Add AbstractGooglePlayServicesChecker.java

This adds an abstract class for the gms independent parts of
GooglePlayServicesChecker.java.

* Add support-v4 compile dependency

This is used in GooglePlayServicesChecker.java

* Add AbstractDynamicStarMapModule.java

This adds a base class for the gms independent parts of
DynamicStarMapModule.java.

* Separate build flavors for gms and fdroid

This adds two build flavors. One using gms (analytics and location) and
one without for F-Droid. All classes using gms are moved to the flavor
source path and a version without gms is added or the fdroid flavor.