Checkpoint: it compiles.
A rather radical restructuring that breaks the dependency of tools on
app. The protocol buffer has been moved into a separate datamodel
project that the other two depend on. Instead of having to run your own
protoc, this is now done for you. Instead of having sed magic to strip
out bits of the .proto file and change names, both clients use the same
version of the file.  The idea is that the tools side will populate the
proto with the string name that should be found in the R file. The app
itself will then populate the integer ids. The code will use a default
of it's missing.

THIS WILL NOT RUN AS IT IS NOW.