More layer code simplification. (#432)
Simplify and Kotlinize even more.  There's still more to do over time, but this should allow work on kotlinizing the renderer to begin and also make it at least possible to add new features.



* Mechanical transformation of the Java layers code to Kotlin.

* Clean up imports etc that were broken during the automatic conversion.

* Simplify AbstractLayer and kotlinize it a bit more.

* Changed to lateinit.

* Remove reduntant code. Reluctantly leave the generics and unsafe casts
as is since eliminating it does add more code, even if it makes it more
readable.

* Kotlinize some more.

* Remove a redundant method.

* Broke a lot of things...

* Fixed stuff I broke.

* There seems to be no need for the closures to be stored in a TreeSet.
The ordering isn't really used. And if that's the case then the Closure
doesn't need to be Comparable.

* Eliminate the UpdateClosure classes and replace with modern functional
stuff.

* Renamed "Source" to "Renderable" which better reflects what they are.

* Mechanical conversion of the renderables to Kotlin and the bare minimum
to get it to build again.

* Post-kotlin conversion cleanup.

* Eliminate another few unnecessary math functions.

* Renamed the source package.

* More renaming.

* Remove an unused method

* Rename a Kotlin-unfriendly variable.

* Upgrade Gradle version.

* Adds a placeholder image of Earth. We don't need it yet and will find a
better one later.
This image is from NASA and therefore in the public domain.

* Honey I shrunk the earth.

* Whoops I didn't mean to commit all that.
Might as well add in these two files: I renamed the Planet enum and its
associated rendered as they're not all planets.

* Revert "Whoops I didn't mean to commit all that."

This reverts commit e7a5b7a0fdc11bf6f2958174d8dabe79b77b1104.

* Revert "Honey I shrunk the earth."

This reverts commit 2491623ff917bdb65587c1b1c0c389fef331f446.

* Honey I shrunk the earth mk 2.

* Update the dd.

* Rename Planet to SolarSystemBody since it represents more than Planets.

* More renaming. The new "Earth" SolarSystem object is included in the
enum so that its orbital elements are available. However this
adds the wrinkle that we now need to remove it from the list of
objects to render.

* Bug fix.

* These are hardly new any more.

* Fix the bug with the planet orientation.
Not sure this code is right, but at least it's the same as before and
doesn't crash.