Add an option to rotate the view 90 degrees for telescopes. (#452)
This change adds a new viewing mode that shows the sky as it is in front of the phone, ie as pointed to by the long edge. The purpose is for telescope users to be able to strap the phone to a telescope's tube.

Took this opportunity to also update the deprecated sensorlistener code to its modern equivalent, which involved a lot of switching around of axes etc to remove the 'fixes' we had to apply to Android's originally buggy sensor code


* Rename some constants to do with the screen rotation so they make more
sense.

* Telescope coordinates.

The view is now as if you're sighting along the edge of the phone
instead of into it.  It's almost right...just the screen seems to be
upside down.

* Update the legacy sensor code to use SensorEventListener instead of the
deprecated SensorListener. This also allows for some sanitization of the
code since we no longer have to deal with various coordinate systems
being backwards.

* Update the docs some more.

* Wired up the 'telescope mode' option.

* Fixed the tests now that the sensor readings have changed.

* Update version number as the Play Store got confused and incorrectly
marked the
previous version number as used.

* Change the name of the new mode and clean up some comments.

* Fix some text in the design doc.

* Change a label to match the casing of the others.